Update Projects, add robocup page
