+++ docs/pages/projects.md
@@ -1,51 +1,58 @@
-# Projects Page
+# Projects
 
-The Projects page provides detailed information about Triton Droids' various robotics projects.
+Projects are split into an overview page and separate project pages.
 
-**Location:** `src/pages/Projects/components/Projects.tsx`  
-**Route:** `/projects`
+**Routes:** `/projects` (index), `/projects/humanoid`, `/projects/learning-projects`, `/projects/robocup`
 
 ## Overview
 
-The Projects page showcases detailed information about the Humanoid Robot project, including:
-
-- **Hero Section** - Title, description, and ASCII art visualization
-- **Detail Cards** - Rationale, Function, Approach, and Progress sections
-- **3D Model Viewer** - Interactive model display (placeholder)
-- **Image Gallery** - Carousel of project images
+- **Index** (`/projects`) – ProjectsSection with three project cards and Learn More links.
+- **Humanoid** (`/projects/humanoid`) – Hero, detail cards, 3D viewer, gallery.
+- **Learning Projects** (`/projects/learning-projects`) – Hero and description (placeholder).
+- **RoboCup** (`/projects/robocup`) – Hero, What is RoboCup and Our Goal cards.
 
 ## Page Structure
 
 ```
 src/pages/Projects/
-  ├── components/
-  │   ├── Projects.tsx           # Main page component
-  │   ├── ProjectHeroSection.tsx # Hero with title and ASCII art
-  │   ├── ProjectDetailCard.tsx  # Reusable card for project details
-  │   ├── ProjectGallery.tsx     # Image carousel using Embla
-  │   └── ModelViewer.tsx        # 3D model viewer placeholder
-  └── ProjectsSection.tsx        # Projects overview (used on Home page)
+  ├── ProjectsIndex.tsx          # /projects
+  ├── Humanoid.tsx               # /projects/humanoid
+  ├── LearningProjects.tsx       # /projects/learning-projects
+  ├── RoboCup.tsx                # /projects/robocup
+  ├── ProjectsSection.tsx        # Overview (Home + index)
+  └── components/
+      ├── Projects.tsx           # Re-exports ProjectsIndex
+      ├── ProjectHeroSection.tsx
+      ├── ProjectDetailCard.tsx
+      ├── ProjectGallery.tsx
+      └── ModelViewer.tsx
 ```
 
 ## Components
 
 ### ProjectHeroSection
+
 Displays the project title, description, and ASCII art asset.
 
 ### ProjectDetailCard
+
 Reusable card component with:
+
 - Icon (88px, centered)
 - Title (64px, gray #A3B2C2)
 - Description (24px, white)
 - Fixed width layout for alignment
 
 ### ProjectGallery
+
 Image carousel using `embla-carousel-react` with:
+
 - Navigation arrows
 - Pagination dots
 - Smooth transitions
 
 ### ModelViewer
+
 Placeholder for 3D model integration (Three.js, model-viewer, or Sketchfab).
 
 ## Related Components
@@ -55,6 +62,7 @@ Placeholder for 3D model integration (Three.js, model-viewer, or Sketchfab).
 ## Assets
 
 Project-related assets are located in:
+
 - `src/assets/projects/` - Project icons and images
   - `humanoid-icon-blue.svg` / `humanoid-icon-gray.svg`
   - `pupper-icon-blue.svg` / `pupper-icon-gray.svg`
@@ -66,4 +74,6 @@ Project-related assets are located in:
   - `icons/function-icon.svg`
   - `icons/approach-icon.svg`
   - `icons/progress-icon.svg`
+  - `icons/what-is-robocup-icon.svg`
+  - `icons/our-goal-icon.svg`
   - `hero/ascii-art.png`
