+++ src/pages/Projects/ProjectsIndex.tsx
@@ -1,0 +1,9 @@
+import ProjectsSection from './ProjectsSection';
+
+export default function ProjectsIndex() {
+  return (
+    <div className="bg-main-bg min-h-screen">
+      <ProjectsSection />
+    </div>
+  );
+}
